clean

+++ examples/opensora_hpcai/opensora/pipelines/infer_pipeline.py
@@ -118,10 +118,10 @@ class InferPipeline:
         # for token/text drop in caption embedder for condition-free guidance training. The null mask is the same as text mask.
         n = x.shape[0]
         # (n_tokens, dim_emb) -> (b n_tokens dim_emb)
-        null_emb = self.model.y_embedder.y_embedding[None, :, :].repeat(n, axis=0)
+        null_emb = self.model.y_embedder.y_embedding[None, :, :].repeat(n, axis=0).to(text_emb.dtype)
 
         if self.use_cfg:
-            y = ops.cat([text_emb, null_emb.to(text_emb.dtype)], axis=0)
+            y = ops.cat([text_emb, null_emb], axis=0)
             x_in = ops.concat([x] * 2, axis=0)
             assert y.shape[0] == x_in.shape[0], "shape mismatch!"
             inputs["mask"] = ops.concat([mask, mask], axis=0)
